chore(refactor): Moves _LobAddressElements into separate module
